projects/prompt-benchmark/requirements.txt
pydantic>=2.5.0
pyyaml>=6.0
aiosqlite>=0.19.0
anthropic>=0.39.0
